Do not pass CMAKE_GENERATOR as variable
From my reading of the documentation:
https://cmake.org/cmake/help/latest/variable/CMAKE_GENERATOR.html this
should only be passed via the -G option, interactively in cmake-gui or
as an environment variable.

If CMAKE_GENERATOR is passed as a variable definition (via. -D) and the
-G option is used as well then cmake gives an error about conflicting
change of generator, as if it expects the CMAKE_GENERATOR variable
definition to specify the generator given a prior cmake run.